% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchLecture.R
\name{searchLecture}
\alias{searchLecture}
\title{search through lecture notes}
\usage{
searchLecture(keywords)
}
\arguments{
\item{keywords}{the desire keywords}
}
\value{
A matrix of the related pages
}
\description{
This function takes one or multiple keywords in and return related materials
and locations. It searches STAT 5293 lecture slides and locate related notes
and codes for query purpose.
The author of STAT 5293 lecture notes is Prof.Joyce Robbins
}
